// src/api/generated/model/groupChallengeRecordFeedResponse.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { GroupDailyVerificationSummaryResponse } from './groupDailyVerificationSummaryResponse';
import type { MemberResponse } from './memberResponse';

export interface GroupChallengeRecordFeedResponse {
  groupId?: number;
  date?: string;
  dailySummary?: GroupDailyVerificationSummaryResponse;
  members?: MemberResponse[];
}