// src/api/generated/model/memberResponse.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { ActivityRecordResponse } from './activityRecordResponse';
import type { MemberDailyGoalResponse } from './memberDailyGoalResponse';
import type { PokedUserResponse } from './pokedUserResponse';
import type { ReactionSummaryResponse } from './reactionSummaryResponse';

export interface MemberResponse {
  groupMemberId?: number;
  groupChallengeParticipantId?: number;
  userId?: number;
  displayName?: string;
  profileImageUrl?: string;
  isUserWithdrawn?: boolean;
  isMe?: boolean;
  memberStatus?: string;
  participantStatus?: string;
  dailyStatus?: string;
  includedInGroupResult?: boolean;
  goals?: MemberDailyGoalResponse[];
  challengeRecordId?: number;
  activityRecord?: ActivityRecordResponse;
  reactionCount?: number;
  commentCount?: number;
  pokeCount?: number;
  isPoked?: boolean;
  pokeable?: boolean;
  pokedUsers?: PokedUserResponse[];
  reactions?: ReactionSummaryResponse;
}